MIC: Factor raw image creation from iso artifacts + move population from intpu iso into the store code base.
