+++ toolkit/tools/pkg/imagecustomizerlib/liveosisoimages.go
@@ -23,6 +23,18 @@ import (
 	"github.com/microsoft/azurelinux/toolkit/tools/internal/targetos"
 )
 
+const (
+	dracutConfig = `add_dracutmodules+=" dmsquash-live livenet selinux "
+add_drivers+=" overlay "
+hostonly="no"
+`
+
+	// the total size of a collection of files is multiplied by the
+	// expansionSafetyFactor to estimate a disk size sufficient to hold those
+	// files.
+	expansionSafetyFactor = 1.5
+)
+
 func createInitrdImage(writeableRootfsDir, kernelVersion, outputInitrdPath string) error {
 	logger.Log.Debugf("Generating initrd (%s) from (%s)", outputInitrdPath, writeableRootfsDir)
